# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediKiosk",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/doctors": {
      "get": {
        "tags": [
          "doctors"
        ],
        "summary": "List Doctors",
        "description": "List available doctors.",
        "operationId": "list_doctors_api_doctors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by department",
              "title": "Department"
            },
            "description": "Filter by department"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by language code",
              "title": "Language"
            },
            "description": "Filter by language code"
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT token for SSE/WebSocket clients",
              "title": "Token"
            },
            "description": "JWT token for SSE/WebSocket clients"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/intake/select-doctor": {
      "post": {
        "tags": [
          "doctors"
        ],
        "summary": "Select Doctor",
        "description": "Select a doctor for a session and persist selection.",
        "operationId": "select_doctor_api_intake_select_doctor_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT token for SSE/WebSocket clients",
              "title": "Token"
            },
            "description": "JWT token for SSE/WebSocket clients"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectDoctorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectDoctorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Generate Token",
        "description": "Issue a signed HS256 JWT token.",
        "operationId": "generate_token_api_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/speech/transcribe": {
      "post": {
        "tags": [
          "speech"
        ],
        "summary": "Transcribe Audio",
        "description": "Transcribe raw audio recording to text.\n\nAccepts raw audio blobs (webm, wav, ogg, etc.) via multipart form-data\nand returns transcribed text.",
        "operationId": "transcribe_audio_api_speech_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_api_speech_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_transcribe_audio_api_speech_transcribe_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Audio recording file (webm/wav)"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Target language code (e.g. 'hi', 'en', 'auto')",
            "default": "auto"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID context"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transcribe_audio_api_speech_transcribe_post"
      },
...
      "DoctorListResponse": {
        "properties": {
          "doctors": {
            "items": {
              "$ref": "#/components/schemas/DoctorResponse"
            },
            "type": "array",
            "title": "Doctors"
          }
        },
        "type": "object",
        "required": [
          "doctors"
        ],
        "title": "DoctorListResponse",
        "description": "API response for a list of doctors."
      },
      "DoctorResponse": {
        "properties": {
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "degrees": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degrees"
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "sub_speciality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Speciality"
          },
          "clinical_interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Clinical Interests"
          },
          "experience_years": {
            "type": "integer",
            "title": "Experience Years"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "seniority_tier": {
            "type": "string",
            "title": "Seniority Tier"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "consultation_fee_inr": {
            "type": "number",
            "title": "Consultation Fee Inr"
          },
          "registration_fee_inr": {
            "type": "number",
            "title": "Registration Fee Inr"
          },
          "followup_free_days": {
            "type": "integer",
            "title": "Followup Free Days"
          },
          "pmjay_accepted": {
            "type": "boolean",
            "title": "Pmjay Accepted"
          },
          "tpa_insurers_accepted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tpa Insurers Accepted"
          },
          "chamber_room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chamber Room"
          },
          "availability_status": {
            "type": "string",
            "title": "Availability Status"
          },
          "opd_start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opd Start Time"
          },
          "opd_end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opd End Time"
          }
        },
        "type": "object",
        "required": [
          "doctor_id",
          "full_name",
          "degrees",
          "department",
          "clinical_interests",
          "experience_years",
          "languages",
          "seniority_tier",
          "review_count",
          "consultation_fee_inr",
          "registration_fee_inr",
          "followup_free_days",
          "pmjay_accepted",
          "tpa_insurers_accepted",
          "availability_status"
        ],
        "title": "DoctorResponse",
        "description": "API response for a single doctor profile."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SelectDoctorRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "doctor_id"
        ],
        "title": "SelectDoctorRequest",
        "description": "Request to select a doctor for a session."
      },
      "SelectDoctorResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "doctor_id",
          "message"
        ],
        "title": "SelectDoctorResponse",
        "description": "Response after selecting a doctor."
      },
...
      "TokenRequest": {
        "properties": {
          "client_id": {
            "type": "string",
            "minLength": 1,
            "title": "Client Id",
            "description": "Client identifier"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role to assume",
            "default": "Kiosk_Device"
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "TokenRequest"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse"
      },
      "TranscribeResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Transcribed speech text"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Detected or processed language code",
            "default": "en"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "description": "Transcription confidence score between 0.0 and 1.0",
            "default": 0.98
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TranscribeResponse",
        "description": "Transcription response payload."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}